make `run_routines()` a standalone function
instead of a method on `struct Program`.

This is so that once the treewalk is done concurrently, `struct Program`
doesn't need to be protected from concurrent mutable access, since it
isn't accessed mutably from `treewalk()`.

`struct ProgramState` will have to be made concurrency-safe, but now
`run_routines()` only needs a mutable reference to `ProgramState` and not
`Program`.