add some error handling, remove some panics
Adds an Error enum with a CompileError (not currently used), an
AttributeInBeginOrEnd error for a program that tries to query a file
attribute in a begin or end block, and an IoError which at this point
can only occur if a stat() fails.

If a stat() fails on a file while the program is running, the program
stops processing that file but keeps processing any remaining files in
the tree.