make global variable assignment atomic
When puffin routines run concurrently, assignment to global variables
might involve an expression on the right hand side that contains the
same or other globals. For example:

    counter = counter + 1

In order to make this read-modify-write atomic, the global variable
mutex is held for the duration of evaluating the RHS expression.

This means that evaluating an expression needs access to the unlocked
globals in some cases, so a new Option parameter is added for that.