new VariableState enum, prep work for associative arrays
In order to make evaluating and assigning variables simpler, a new
`enum VariableState` is created which can hold either a locked variables
vector or a reference to an unlocked variables vector.

Methods for evaluating expressions, identifiers, etc. can just take this
enum and thus don't have to know whether they're being evaluated in a
context where the variables are locked or unlocked.