associative arrays implementation: get array values
This adds support for getting values in associative arrays to the AST.

Setting values is not yet supported, nor is parsing associative arrays.

The `Assignment` struct is updated to contain a `Variable` on the left
hand side, and not just an `Identifier`, so that both types of variables
can be assigned to.