add analysis pass to resolve variable types
This adds an analysis pass to walk the AST and convert
`Variable::NotYetKnown` objects to the appropriate kind of Variable.

Whether an identifier refers to an array or scalar is inferred from its
use. If there is at least one use where it is followed by a subscript
expression, it is treated as an array. Otherwise, it is treated as a
scalar.

Since the analysis pass needs knowledge of the AST structure, it is
moved into a submodule of `mod ast`.

There is one kind of type error now: using an array name in a context
where an array name cannot be used (for example on the left-hand side of
an assignment). This error is found at compile-time, so for example the
following program will not compile:

    # Error:
    puffin . '{ arr = 1; arr["sub"] = 1 }'