add tests for arrays and variables; fix array bug
This adds tests for arrays and variables and fixes a bug where
evaluating an array subscript expression would be done on the Locked
VariableState, but the VariableState was already locked to lock the
array itself, so the program would deadlock. The solution is to evaluate
array subscript expressions using VariableState::Unlocked.