new Value types for special types like inodes, devices numbers, etc.
A number of types are represented as integers despite not being
integers semantically because it makes no sense to do arithmetic on
them. This includes many "IDs" that filesystems use, such as inode and
device numbers, UIDs, and GIDs, etc.

Such types are convenient to represent as integers but it shouldn't be
allowed to use them in contexts where a number is expected, such as an
arithmetic expression. What would it mean to add inode numbers, or to
multiply a UID by 2? This makes no sense, so make it an error in the
language.